requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
sendgrid>=6.11.0
asyncio>=3.4.3
